% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_study_variables.R
\name{check_study_variables}
\alias{check_study_variables}
\title{Compare variable consistency in two csv files - exact match}
\usage{
check_study_variables(
  path_study_vars_reference,
  ref_colname,
  path_metadata,
  metadata_colname
)
}
\arguments{
\item{path_study_vars_reference}{file path of reference meta data file (.csv)
e.g. "p_meta_data/PfizerRF_study_variables.csv"}

\item{ref_colname}{column name to be checked from the reference meta data csv
e.g. "VarName"}

\item{path_metadata}{file path of meta data file of interest (.csv)
e.g. "Pfizer_algorithms.csv"}

\item{metadata_colname}{column name to be checked from the meta data csv of interest
e.g. "CONCEPT"}
}
\value{
a list of "not found" variables
}
\description{
Given the paths of two csv files (";" separated) and the column names,
this function returns the list of "not found" variables that are not
present in the reference file (first submitted file), but present in the second
submitted file
}
