% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadatachecker_app.R
\name{metadatachecker_app}
\alias{metadatachecker_app}
\title{metadatachecker shiny app}
\usage{
metadatachecker_app(...)
}
\arguments{
\item{...}{no argument is supplied}
}
\value{
Opens the shiny app in a new window or in browser
}
\description{
In the console, type: metadatachecker_app()
}
